removing lr_qaoa_helpers
